Add files via upload
Viterbi Algorithm is the main algorithm for Most Likely Explanation (MLE) in HMM.  We study the hypothesis that the prediction accuracy of Viterbi algorithm can be estimated a prior by computing the arithmetic mean of the cosines of the emission probabilities.  Our analysis and experimental results suggest a close relationship between these two quantities.